st：更新业务板块-根据code

+++ src/main/java/ttjj/dto/RankBizDataDiff.java
@@ -168,7 +168,7 @@ public class RankBizDataDiff {
     /**
      * f62 今日主力净流入
      */
-    private Long f62;
+    private BigDecimal f62;
     private Double f63;
     private Double f64;
     private Double f65;
@@ -792,11 +792,11 @@ public class RankBizDataDiff {
         this.f61 = f61;
     }
 
-    public Long getF62() {
+    public BigDecimal getF62() {
         return f62;
     }
 
-    public void setF62(Long f62) {
+    public void setF62(BigDecimal f62) {
         this.f62 = f62;
     }
